serialize mapped_column_ids to vector<idx_t> instead of vector<StorageIndex>
