+++ src/include/duckdb/storage/serialization/storage.json
@@ -81,27 +81,8 @@
     "pointer_type": "none"
   },
   {
-    "class": "StorageIndex",
-    "members": [
-      {
-        "id": 100,
-        "name": "index",
-        "type": "idx_t"
-      },
-      {
-        "id": 101,
-        "name": "child_indexes",
-        "type": "vector<StorageIndex>"
-      }
-    ],
-    "pointer_type": "none"
-  },
-  {
     "class": "IndexStorageInfo",
-    "includes": [
-      "duckdb/storage/table_storage_info.hpp",
-      "duckdb/storage/storage_index.hpp"
-    ],
+    "custom_implementation": true,
     "members": [
       {
         "id": 100,
@@ -127,7 +108,7 @@
       {
         "id": 104,
         "name": "mapped_column_ids",
-        "type": "vector<StorageIndex>"
+        "type": "vector<idx_t>"
       },
       {
         "id": 105,
